requirements.txt
google-generativeai>=0.8.0
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
gtts>=2.5.0
pillow>=10.0.0

